+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Added `Array::retrieve_chunk_if_exists` and variants (`async_`, `_elements`, `_ndarray`)
  - Implement `AsyncBytesPartialDecoderTraits` for `std::io::Cursor<{&[u8],Vec<u8>}>`
  - Added missing partial decoding tests for codecs
+ - Implement `From<ChunkShape>` for `Vec<NonZeroU64>`
 
 ### Changed
  - Dependency bumps